refactor: move db model to separate module
